Change type of baud argument in Dynamixel_SetBaudRate from long to uint32_t
It is better to stick with platform-independent types, and 32 bits is sufficient for the baud argument.


Former-commit-id: b03709166dd1318ae915e715a39c022f2b51a6a3 [formerly 0f926c685c593e89ae57edfeec9b9b317ce89563]
Former-commit-id: 3897f67e2e33f8593da5f47f22817c73e809878a

+++ Robot/Drivers/Dynamixel/DynamixelProtocolV1.c
@@ -99,7 +99,7 @@ void Dynamixel_SetID(Dynamixel_HandleTypeDef* hdynamixel, uint8_t ID){
 	hdynamixel -> _ID = ID;
 }
 
-void Dynamixel_SetBaudRate(Dynamixel_HandleTypeDef* hdynamixel, long baud){
+void Dynamixel_SetBaudRate(Dynamixel_HandleTypeDef* hdynamixel, uint32_t baud){
 	/* Sets the baud rate of a particular motor. Register address is 0x04 in motor EEPROM.
 	 *
 	 * Instruction register address: 0x04 (EEPROM)
